Updated Version Handling + Cleaned Up Output
In order to support specifying depdencies with versions, the pkgName
and version needed to be able to be supplied in a single string. To
this end, I've adopted VALET's method of pkgName/pkgVersion. It seems
to be working so far.

Also, some effort was put into making the output of typyEnv (especially
when it errors/fails), a little cleaner.